prototype/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
openai==0.27.4
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pycparser==2.21
python-dateutil==2.8.2
python-docx==0.8.11
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
tenacity==8.2.2
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.65.0
urllib3==1.26.15
Wand==0.6.11
Werkzeug==2.2.3
yarl==1.9.1